Clean up buttons toolbar code slightly
This splits up `buttons.h` a bit and unnests various classes.
