+++ foo_ui_columns/buttons.h
@@ -4,233 +4,51 @@
 
 #include "core_dark_list_view.h"
 
+#include "buttons_button.h"
+
 namespace cui::toolbars::buttons {
+
+enum class IconSize : int32_t {
+    Automatic,
+    Custom,
+};
+
+enum Appearance : uint32_t {
+    APPEARANCE_NORMAL,
+    APPEARANCE_FLAT,
+    APPEARANCE_NOEDGE,
+};
+
+enum Identifier {
+    I_TEXT_BELOW,
+    I_APPEARANCE,
+    I_BUTTONS,
+    I_ICON_SIZE,
+    I_WIDTH,
+    I_HEIGHT,
+};
+
+enum ButtonIdentifier {
+    I_BUTTON_TYPE,
+    I_BUTTON_FILTER,
+    I_BUTTON_SHOW,
+    I_BUTTON_GUID,
+    I_BUTTON_CUSTOM,
+    I_BUTTON_CUSTOM_DATA,
+    I_BUTTON_CUSTOM_HOT,
+    I_BUTTON_CUSTOM_HOT_DATA,
+    I_BUTTON_RESERVED_1,
+    I_BUTTON_RESERVED_2,
+    I_BUTTON_RESERVED_3,
+    I_BUTTON_RESERVED_4,
+    I_BUTTON_USE_CUSTOM_TEXT,
+    I_BUTTON_TEXT,
+    I_BUTTON_SUBCOMMAND
+};
 
 class ButtonsToolbar : public uie::container_uie_window_v3 {
 public:
-    enum class ConfigVersion {
-        VERSION_1,
-        VERSION_2,
-        VERSION_CURRENT = VERSION_2
-    };
-    enum class FCBVersion {
-        VERSION_1,
-        VERSION_2,
-        VERSION_3,
-        VERSION_CURRENT = VERSION_3
-    };
-
-    /** For config dialog */
-    enum {
-        MSG_BUTTON_CHANGE = WM_USER + 2,
-        MSG_COMMAND_CHANGE = WM_USER + 3
-    };
-
-    enum class IconSize : int32_t {
-        Automatic,
-        Custom,
-    };
-
-    enum Filter : int32_t {
-        FILTER_NONE,
-        FILTER_PLAYING,
-        FILTER_PLAYLIST,
-        FILTER_ACTIVE_SELECTION,
-    };
-
-    enum Type : int32_t {
-        TYPE_SEPARATOR,
-        TYPE_BUTTON,
-        TYPE_MENU_ITEM_CONTEXT,
-        TYPE_MENU_ITEM_MAIN,
-    };
-
-    enum Show : uint32_t {
-        SHOW_IMAGE,
-        SHOW_IMAGE_TEXT,
-        SHOW_TEXT,
-    };
-
-    enum Appearance : uint32_t {
-        APPEARANCE_NORMAL,
-        APPEARANCE_FLAT,
-        APPEARANCE_NOEDGE,
-    };
-
-    enum Identifier {
-        I_TEXT_BELOW,
-        I_APPEARANCE,
-        I_BUTTONS,
-        I_ICON_SIZE,
-        I_WIDTH,
-        I_HEIGHT,
-    };
-    enum ButtonIdentifier {
-        I_BUTTON_TYPE,
-        I_BUTTON_FILTER,
-        I_BUTTON_SHOW,
-        I_BUTTON_GUID,
-        I_BUTTON_CUSTOM,
-        I_BUTTON_CUSTOM_DATA,
-        I_BUTTON_CUSTOM_HOT,
-        I_BUTTON_CUSTOM_HOT_DATA,
-        I_BUTTON_RESERVED_1,
-        I_BUTTON_RESERVED_2,
-        I_BUTTON_RESERVED_3,
-        I_BUTTON_RESERVED_4,
-        I_BUTTON_USE_CUSTOM_TEXT,
-        I_BUTTON_TEXT,
-        I_BUTTON_SUBCOMMAND
-    };
-
-    enum CustomImageIdentifier {
-        I_CUSTOM_BUTTON_PATH,
-        I_CUSTOM_BUTTON_MASK_PATH,
-        I_CUSTOM_BUTTON_MASK_TYPE = 8,
-        I_CUSTOM_BUTTON_IMAGE_DATA = 9,
-        I_CUSTOM_BUTTON_MASK_DATA = 10,
-        I_CUSTOM_BUTTON_MASK_COLOUR = 11
-    };
-
-    enum ImageIdentifier {
-        IMAGE_NAME,
-        IMAGE_DATA,
-        IMAGE_PATH
-    };
-
-    enum class CustomImageContentType {
-        Ico,
-        Svg,
-        Other,
-    };
-
-    class ButtonStateCallback : public uie::button_callback {
-    public:
-        ButtonStateCallback(uie::button::ptr button_instance, HWND toolbar_wnd, int button_id)
-            : m_button_instance(std::move(button_instance))
-            , m_toolbar_wnd(toolbar_wnd)
-            , m_button_id(button_id)
-        {
-            m_button_instance->register_callback(*this);
-        }
-
-        ~ButtonStateCallback() { m_button_instance->deregister_callback(*this); }
-
-    private:
-        void on_button_state_change(unsigned p_new_state) final;
-
-        void on_command_state_change(unsigned p_new_state) final {}
-
-        uie::button::ptr m_button_instance;
-        HWND m_toolbar_wnd{};
-        int m_button_id{};
-    };
-
-    class MainMenuStateCallback : public mainmenu_commands_v3::state_callback {
-    public:
-        MainMenuStateCallback(mainmenu_commands_v3::ptr mainmenu_commands_v3_instance, uint32_t command_index,
-            HWND toolbar_wnd, int button_id)
-            : m_mainmenu_commands_v3(std::move(mainmenu_commands_v3_instance))
-            , m_command_index(command_index)
-            , m_toolbar_wnd(toolbar_wnd)
-            , m_button_id(button_id)
-        {
-            m_mainmenu_commands_v3->add_state_callback(this);
-        }
-
-        ~MainMenuStateCallback() { m_mainmenu_commands_v3->remove_state_callback(this); }
-
-    private:
-        void menu_state_changed(const GUID& main, const GUID& sub) final;
-
-        mainmenu_commands_v3::ptr m_mainmenu_commands_v3;
-        uint32_t m_command_index{};
-        HWND m_toolbar_wnd{};
-        int m_button_id{};
-    };
-
-    class Button {
-    public:
-        class CustomImage {
-        public:
-            pfc::string_simple m_path;
-            pfc::string_simple m_mask_path;
-            COLORREF m_mask_colour{};
-            ui_extension::t_mask m_mask_type{uie::MASK_NONE};
-
-            [[nodiscard]] CustomImageContentType content_type() const;
-            [[nodiscard]] pfc::string8 get_path() const;
-
-            void write(stream_writer* out, abort_callback& p_abort) const;
-            void read(ConfigVersion p_version, stream_reader* reader, abort_callback& p_abort);
-            void write_to_file(stream_writer& p_file, bool b_paths, abort_callback& p_abort);
-            void read_from_file(FCBVersion p_version, const char* p_base, const char* p_name, stream_reader* p_file,
-                unsigned p_size, abort_callback& p_abort);
-        };
-
-        Type m_type{TYPE_SEPARATOR};
-        Filter m_filter{FILTER_ACTIVE_SELECTION};
-        Show m_show{SHOW_IMAGE};
-        GUID m_guid{};
-        GUID m_subcommand{};
-        bool m_use_custom{false};
-        bool m_use_custom_hot{false};
-        bool m_use_custom_text{false};
-        pfc::string_simple m_text;
-        CustomImage m_custom_image;
-        CustomImage m_custom_hot_image;
-
-        uie::button::ptr m_interface;
-        mainmenu_commands_v3::ptr m_mainmenu_commands_v3;
-        std::optional<uint32_t> m_mainmenu_commands_index{};
-        std::shared_ptr<ButtonStateCallback> m_button_state_callback;
-        std::shared_ptr<MainMenuStateCallback> m_mainmenu_state_callback;
-
-        void reset_state();
-
-        void write(stream_writer* out, abort_callback& p_abort) const;
-
-        void read(ConfigVersion p_version, stream_reader* reader, abort_callback& p_abort);
-        std::string get_display_text() const;
-
-        std::string get_type_desc() const;
-        std::string get_name(bool short_form = false) const;
-        std::string get_name_with_type() const;
-        void write_to_file(stream_writer& p_file, bool b_paths, abort_callback& p_abort);
-        void read_from_file(FCBVersion p_version, const char* p_base, const char* p_name, stream_reader* p_file,
-            unsigned p_size, abort_callback& p_abort);
-    };
-
-    class ButtonImage {
-    public:
-        ButtonImage() = default;
-        ButtonImage(const ButtonImage&) = delete;
-        ButtonImage& operator=(const ButtonImage&) = delete;
-        ButtonImage(ButtonImage&&) = delete;
-        ButtonImage& operator=(ButtonImage&&) = delete;
-
-        bool is_valid() const;
-        void preload(const Button::CustomImage& p_image);
-        bool load(std::optional<std::reference_wrapper<Button::CustomImage>> custom_image,
-            const service_ptr_t<uie::button>& button_ptr, COLORREF colour_btnface, int width, int height);
-        unsigned add_to_imagelist(HIMAGELIST iml);
-        [[nodiscard]] std::optional<std::tuple<int, int>> get_size() const { return m_bitmap_source_size; }
-
-    private:
-        bool load_custom_image(const Button::CustomImage& custom_image, int width, int height);
-        void load_custom_svg_image(const char* full_path, int width, int height);
-        void load_default_image(
-            const service_ptr_t<uie::button>& button_ptr, COLORREF colour_btnface, int width, int height);
-
-        wil::com_ptr<IWICBitmapSource> m_bitmap_source;
-        std::optional<std::tuple<int, int>> m_bitmap_source_size;
-
-        wil::unique_hbitmap m_bm;
-        wil::unique_hicon m_icon;
-        ui_extension::t_mask m_mask_type{uie::MASK_NONE};
-        wil::unique_hbitmap m_bm_mask;
-        COLORREF m_mask_colour{0};
-    };
+    static constexpr GUID extension_guid{0xd8e65660, 0x64ed, 0x42e7, {0x85, 0xb, 0x31, 0xd8, 0x28, 0xc2, 0x52, 0x94}};
 
     HWND wnd_toolbar{nullptr};
     HWND wnd_host{nullptr};
@@ -239,8 +57,6 @@ public:
 
     ButtonsToolbar();
     ~ButtonsToolbar();
-
-    static const GUID extension_guid;
 
     const GUID& get_extension_guid() const override;
 
@@ -337,13 +153,11 @@ public:
 private:
     uie::container_window_v3_config get_window_config() override
     {
-        uie::container_window_v3_config config(class_name);
+        uie::container_window_v3_config config(L"columns_ui_buttons_toolbar_XXpeF2WDs-4");
         config.forward_wm_settingchange = false;
         config.invalidate_children_on_move_or_resize = true;
         return config;
     }
-
-    static const TCHAR* class_name;
 
     WNDPROC menuproc{nullptr};
     bool initialised{false};
@@ -361,8 +175,8 @@ private:
 struct CommandPickerData {
     GUID guid{};
     GUID subcommand{};
-    int group{ButtonsToolbar::TYPE_SEPARATOR};
-    int filter{ButtonsToolbar::FILTER_ACTIVE_SELECTION};
+    int group{TYPE_SEPARATOR};
+    int filter{FILTER_ACTIVE_SELECTION};
 };
 
 class CommandPickerDialog {
@@ -400,24 +214,14 @@ private:
 } // namespace cui::toolbars::buttons
 
 namespace pfc {
+
 template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::ImageIdentifier> : public traits_rawobject {};
+class traits_t<cui::toolbars::buttons::ButtonIdentifier> : public traits_rawobject {};
 template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::CustomImageIdentifier> : public traits_rawobject {};
+class traits_t<cui::toolbars::buttons::Identifier> : public traits_rawobject {};
 template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::ButtonIdentifier> : public traits_rawobject {};
-template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::Identifier> : public traits_rawobject {};
-template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::Appearance> : public traits_rawobject {};
-template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::ConfigVersion> : public traits_rawobject {};
-template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::Show> : public traits_rawobject {};
-template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::Filter> : public traits_rawobject {};
-template <>
-class traits_t<cui::toolbars::buttons::ButtonsToolbar::Type> : public traits_rawobject {};
+class traits_t<cui::toolbars::buttons::Appearance> : public traits_rawobject {};
 template <>
 class traits_t<uie::t_mask> : public traits_rawobject {};
+
 } // namespace pfc
